+++ Source/Components/GAMs/SimulinkWrapperGAM/SimulinkWrapperGAM.cpp
@@ -1692,7 +1692,7 @@ bool SimulinkWrapperGAM::ScanSignal(const uint32 sigIdx, StreamString spacer, co
     uint32              ELEdimArrayIdx;
     rtwCAPI_Orientation ELEorientation;
     
-    uint32              ELEsize         = 1u;
+    uint64              ELEsize         = 1u;
     uint32              ELEMARTeNumDims = 0u;   // number of dimensions according to MARTe standard
     
     StreamString        fullPathName;
